+++ src/parse_args.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "util/expected.hpp"
+#include <optional>
 #include <stdexcept>
 #include <string>
 #include <vector>
@@ -18,5 +19,13 @@ struct mount_entry {
 /// mountpoint
 util::expected<std::vector<mount_entry>, std::runtime_error>
 parse_arg(const std::string &arg, std::optional<std::string> uenv_repo_path);
+
+struct uenv_desc {
+  using entry_t = std::optional<std::string>;
+  entry_t name{std::nullopt};
+  entry_t version{std::nullopt};
+  entry_t tag{std::nullopt};
+  entry_t sha{std::nullopt};
+};
 
 } // namespace impl
