+++ src/util/helper.hpp
@@ -1,7 +1,7 @@
 #include <string>
 #include <vector>
 
-std::vector<std::string> split(const std::string &s, char delim, int maxsplit=-1);
+std::vector<std::string> split(const std::string &s, char delim);
 
 bool is_sha(const std::string &str);
